search user by search term

+++ routes/userRoutes.js
@@ -36,6 +36,22 @@ userRouter.get('/:userId', (req, res, next) => {
 	})
 })
 
+// get user(s) by search term
+userRouter.get('/search', (req, res, next) => {
+	// search term
+	const { user } = req.query
+	// creates a regular expression out of the string "user"
+	const pattern = new RegExp(user) // -- /user/
+	//$options: 'i'  Case insensitive -lower or uppercase
+	User.find({ name: { $regex: pattern, $options: 'i' } }, (err, users) => {
+		if (err) {
+			res.status(500)
+			return next(err)
+		}
+		return res.status(200).send(users)
+	})
+})
+
 // update user
 userRouter.put('/:userId', (req, res, next) => {
 	User.findOneAndUpdate(
@@ -54,20 +70,15 @@ userRouter.put('/:userId', (req, res, next) => {
 
 // DELETE user
 userRouter.delete('/:userId', (req, res, next) => {
-	User.findOneAndDelete(
-		{ _id: req.params.userId },
-		(err, deletedUser) => {
-			if (err) {
-				res.status(500)
-				return next(err)
-			}
-			return res
-				.status(200)
-				.send(
-					`Successfully deleted ${deletedUser.email} from the database`
-				)
+	User.findOneAndDelete({ _id: req.params.userId }, (err, deletedUser) => {
+		if (err) {
+			res.status(500)
+			return next(err)
 		}
-	)
+		return res
+			.status(200)
+			.send(`Successfully deleted ${deletedUser.email} from the database`)
+	})
 })
 
 module.exports = userRouter
